Updated jribble library to latest version.
Updated jribble library to latest version
that has a bit improved caching. Due to
the fact that this version requires
java.io.Reader passed to it to support
reset() method we had to introduce
a little anonymous class acting as a
wrapper for Readers that do not support
reset themselves.

Change-Id: Ia75cca4eeaaed4c93fd887a8418e5bd66ce7ba35
Signed-off-by: Grzegorz Kossakowski <grzegorz.kossakowski@gmail.com>